feat: strategy-armory-rule-weight，抽奖前置规则过滤

+++ big-market-infrastructure/src/main/java/cn/chaZ/infrastructure/persistent/po/StrategyAward.java
@@ -16,7 +16,7 @@ public class StrategyAward {
     /** 自增ID */
     private Long id;
     /** 抽奖策略ID */
-    private Integer strategyId;
+    private Long strategyId;
     /** 抽奖奖品ID */
     private Integer awardId;
     /** 抽奖奖品标题 */
